moved starCalc function into separate file
